Enhanced ContentType Parsing and Referencing (#132)
Fixes #129

- Simplified ContentType.FromString parsing
- Made the raw ContentType string value available to the server Request and
  Response via their respective Advanced.ContentType. Can be directly set on
  the Response if there is not a value in the Enum that matches.
- Added CUSTOM_TEXT and CUSTOM_BINARY to the ContentType enum

To set a custom content type value, set Response.ContentType to either
CUSTOM_TEXT or CUSTOM_BINARY (depending on whether you want to send a text
or binary response), and then set Response.Advanced.ContentType to the
custom value. It must be done in this order, because setting the
Response.ContentType value writes directly to
Response.Advanced.ContentType.