Fix issue 205 and include fixes for issues 202 and 203 (#206)
* Add check to catch() in Route(object state) to not send the exception
message-body in the case of HttpMethod.HEAD. This makes Grapevine
function within the HEAD spec stating that no message-body should be
sent by the server. Otherwise, a System.Net.ProtocolViolationException
will be thrown when the response is sent.

* Add check to see if exception is null. If so, use string.Empty as the
message body. This prevents an exception in the case that
Router.SendExceptionMessages = false.

* Change equality operator to fix broken tests.